+++ tests/Integration/Controller/FrontpageControllerTest.php
@@ -2,8 +2,6 @@
 
 namespace App\Tests\Integration\Controller;
 
-use App\Entity\User;
-use App\Repository\UserRepository;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
 class FrontpageControllerTest extends WebTestCase