+++ tests/Unit/Service/BillingServiceTest.php
@@ -16,7 +16,6 @@ use App\Model\Invoices\ConfirmData;
 use App\Repository\InvoiceEntryRepository;
 use App\Repository\InvoiceRepository;
 use App\Service\BillingService;
-use Doctrine\Common\Collections\ArrayCollection;
 use PHPUnit\Framework\TestCase;
 use Symfony\Contracts\Translation\TranslatorInterface;